requirements.txt
huggingface-hub==0.27.1
lancedb==0.18.0
langchain
langchain-community
langchain-core
matplotlib==3.10.0
numpy
openai-whisper
opencv-python
pandas
pillow
python-dotenv==1.0.1
tiktoken==0.8.0
tokenizers==0.21.0
torch
torchaudio
torchvision
tqdm==4.67.1
transformers==4.48.0
umap-learn
webvtt-py==0.5.1
youtube-transcript-api==0.6.3
